backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]==1.7.4
python-dotenv
alembic
httpx
python-multipart
pydantic>=2.0,<3.0
pydantic-settings
email-validator
fastapi-mail
boto3
jinja2

# Google OAuth
google-auth
google-auth-oauthlib
google-auth-httplib2

# Replicate API
replicate

# Optional performance boost
orjson

